--- DemoApp/Sources/Extensions/CallStateResponseFields+Identifiable.swift
@@ -1,12 +1,0 @@
-//
-// Copyright © 2025 Stream.io Inc. All rights reserved.
-//
-
-import Foundation
-import StreamVideo
-
-extension CallStateResponseFields: Identifiable {
-    public var id: String {
-        call.cid
-    }
-}
