+++ DemoApp/Sources/Views/CallView/CallingView/SimpleCallingView.swift
@@ -243,9 +243,3 @@ struct SimpleCallingView: View {
         }
     }
 }
-
-extension URL: Identifiable {
-    public var id: ObjectIdentifier {
-        .init(absoluteString as NSString)
-    }
-}
